requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
praw==7.8.1
prawcore==2.4.0
requests==2.29.0
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
pycrypto==2.6.1
pycryptodome==3.21.0